GameScreen now listens to GameManager for map changes
